refactor(lab/17): shuffling things around more
update calls stop, but it's not the stop that comes back from play state ... what is it?
