+++ src/lab/17-collisions/src/util/env.ts
