+++ src/lab/17-collisions/src/util/loop.ts
