+++ src/lab/17-collisions/src/util/math.ts
@@ -1,0 +1,4 @@
+import { lerp, random } from '../../../../lib/core/math';
+
+export const rng: (a: number, b: number) => number = (a, b) =>
+  lerp(a, b, random());
